Remove extra linting disable commands and add docstrings
