+++ accounts-journey/configuration/Configuration.swift
@@ -18,12 +18,10 @@ extension AccountsJourney {
         /// Configuration for ``AccountsList`` screen
         public var accountsList = AccountsList.Configuration()
         
+        /// Configuration for ``accountDetails`` screen
         public var accountDetails = AccountDetails.Configuration()
         
         /// Configuration of styles used in the Accounts Journey
         public var design = Design()
-        
-        /// Strings used in the Accounts Journey
-        public var strings = Strings()
     }
 }
