+++ accounts-journey/configuration/Styles.swift
@@ -35,6 +35,7 @@ extension AccountsJourney.Design {
             navigationBar.prefersLargeTitles = true
         }
         
+        /// Journey currency formatter
         public var currencyFormatter: StyleSelector<(String, String), UILabel> = { values in
             return { label in
                 let (amount, currencyCode) = values
@@ -44,8 +45,9 @@ extension AccountsJourney.Design {
                 amountOptions.minFractionDigits = DesignSystem.Formatting.defaultFractionDigits
                 amountOptions.maxFractionDigits = DesignSystem.Formatting.defaultFractionDigits
                 
-                let formatter = DesignSystem.Formatting.makeFormatter()
-               let numberLabel = formatter.format(amount: Decimal(string: amount) ?? 0.00, options: amountOptions)
+                let formatter = DesignSystem.shared.formatting.amountFormatter
+                
+                let numberLabel = formatter.format(amount: Decimal(string: amount) ?? 0.00, options: amountOptions)
                 label.text = numberLabel
                 label.accessibilityLabel = numberLabel
             }
