+++ accounts-journey/configuration/account_details/AccountDetailsConfiguration.swift
