+++ accounts-journey/configuration/account_details/AccountDetailsDesign.swift
