+++ accounts-journey/configuration/account_details/AccountDetailsStrings.swift
