+++ accounts-journey/configuration/accounts_list/AccountListRouter.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 extension AccountsList {
-    
     /// Accounts List router
     public struct Router {
         var didSelectProduct: (UINavigationController) -> (String) -> Void
