+++ accounts-journey/configuration/accounts_list/AccountsListError.swift
@@ -1,56 +1,57 @@
 //
-//  AccountsDetailsError.swift
+//  AccountsListError.swift
 //  AccountsJourney
 //
-//  Created by Backbase R&D B.V. on 28/11/2023.
+//  Created by Backbase R&D B.V. on 27/11/2023.
 //
 
 import Foundation
 import Resolver
 import ClientCommonGen2
 
-extension AccountDetails {
-    /// Generic ErrorFormatter
-    public typealias ErrorFormatter<T> = MainErrorFormatter<AccountDetails.Error, T>
+extension AccountsList {
+    /// Account list ErrorFormatter
+    public typealias ErrorFormatter<T> = MainErrorFormatter<AccountsList.Error, T>
     
-    /// Possible errors
+    /// Accounts lists possible errors
     public enum Error: Swift.Error, Equatable {
         case invalidResponse
         case notConnected
+        case noAccounts
         case loadingFailure(_ response: AccountsJourney.ErrorResponse? = nil)
     }
     
-    static func getError(_ error: Swift.Error, response: AccountsJourney.ErrorResponse?) -> AccountDetails.Error {
+    static func getError(_ error: Swift.Error, response: AccountsJourney.ErrorResponse?) -> AccountsList.Error {
         switch error as? ClientCommonGen2.CallError {
         case .nilHTTPResponse:
-            return AccountDetails.Error.notConnected
+            return AccountsList.Error.notConnected
         default:
-            return AccountDetails.Error.loadingFailure(response)
+            return AccountsList.Error.loadingFailure(response)
         }
     }
 }
 
-extension AccountDetails.Error {
+extension AccountsList.Error {
     /// Name of the icon
     public var icon: String? {
         let config: AccountsJourney.Configuration = Resolver.resolve()
-        return config.accountDetails.design.errorIconName(self)
+        return config.accountsList.design.errorIconName(self)
     }
     /// Title for the given error
     public var title: String {
         let config: AccountsJourney.Configuration = Resolver.resolve()
-        return config.accountDetails.strings.errorTitle(self)()
+        return config.accountsList.strings.errorTitle(self)()
     }
     
     /// Subtitle for the given error
     public var subtitle: String {
         let config: AccountsJourney.Configuration = Resolver.resolve()
-        return config.accountDetails.strings.errorSubtitle(self)()
+        return config.accountsList.strings.errorSubtitle(self)()
     }
     
     /// Action button title for the given error
     public var actionButtonTitle: String? {
         let config: AccountsJourney.Configuration = Resolver.resolve()
-        return config.accountDetails.strings.errorRetryButtonTitle(self)?()
+        return config.accountsList.strings.errorRetryButtonTitle(self)?()
     }
 }
