+++ accounts-journey/domain/model/account_details/AccountDetailsModel.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 extension AccountsJourney {
-    // swiftlint:disable line_length type_body_length function_body_length superfluous_disable_command
+    // swiftlint:disable function_body_length type_body_length
     public struct AccountDetailsModel {
         public let id: String
         public let productId: String
@@ -263,5 +263,5 @@ extension AccountsJourney {
                 self.externalAccountStatus = externalAccountStatus
             }
     }
-// swiftlint:enable line_length type_body_length function_body_length superfluous_disable_command
+// swiftlint:enable function_body_length type_body_length
 }
