+++ accounts-journey/domain/model/account_summary/custom-product/GeneralAccount.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 extension AccountsJourney {
-    // swiftlint:disable line_length type_body_length function_body_length superfluous_disable_command
+    // swiftlint:disable function_body_length
     /// General account type.
     public struct GeneralAccount: Equatable {
         /// Booked balance
@@ -425,5 +425,5 @@ extension AccountsJourney {
         }
        
     }
-    // swiftlint:enable line_length type_body_length function_body_length superfluous_disable_command
+    // swiftlint:enable function_body_length
 }
