+++ accounts-use-case/Mappers/account_detail/ArrangementItemMapper.swift
@@ -8,7 +8,7 @@
 import AccountsJourney
 import ArrangementsClient2Gen2
 
-// swiftlint:disable line_length function_body_length superfluous_disable_command
+// swiftlint:disable function_body_length
 extension ArrangementsClient2Gen2.AccountArrangementItem {
     public func toDomainModel() -> AccountsJourney.AccountDetailsModel {
         AccountsJourney.AccountDetailsModel(
@@ -98,4 +98,4 @@ extension ArrangementsClient2Gen2.AccountArrangementItem {
         )
     }
 }
-// swiftlint:enable line_length  function_body_length superfluous_disable_command
+// swiftlint:enable function_body_length
