+++ accounts-use-case/Mappers/accounts_summary/LoanMapper.swift
@@ -10,7 +10,7 @@ import AccountsJourney
 import ArrangementsClient2Gen2
 
 extension ArrangementsClient2Gen2.Loan {
-    // swiftlint:disable line_length type_body_length function_body_length superfluous_disable_command
+    // swiftlint:disable function_body_length
     public func toDomainModel() -> AccountsJourney.Loan {
         AccountsJourney.Loan(
             bookedBalance: self.bookedBalance,
@@ -64,7 +64,7 @@ extension ArrangementsClient2Gen2.Loan {
             bankBranchCode: self.bankBranchCode2
         )
     }
-    // swiftlint:enable line_length type_body_length function_body_length superfluous_disable_command
+    // swiftlint:enable function_body_length
 }
 
 extension ArrangementsClient2Gen2.LoanProductKinds {
